MAJOR reorganization to break up source code into multiple files
