+++ hiccup/hiccup_constants.py
@@ -1,0 +1,3 @@
+# constant parameter values for HICCUP calculations
+
+tk_zero = 273.15 # value for converting between celsius and Kelvin
