+++ test_scripts/test.SST_from_NOAA.py
@@ -5,8 +5,7 @@
 # temperature and sea ice data files from NOAA data
 # ==================================================================================================
 import os
-from hiccup import hiccup_data_class as hdc
-from hiccup import hiccup_state_adjustment as hsa
+from hiccup import hiccup
 # ------------------------------------------------------------------------------
 
 # local path for grid and mapping files (move this a scratch space for large grids)
@@ -24,28 +23,22 @@ output_sst_file_name = f'{data_tmp}/HICCUP_TEST_OUTPUT.NOAA.sstice.{dst_horz_gri
 # Create data class instance, which includes xarray file dataset objects
 # and variable name dictionaries for mapping between naming conventions.
 # This also checks input files for required variables
-hiccup_data = hdc.create_hiccup_data(name='NOAA'
-                                    ,dst_horz_grid=dst_horz_grid
-                                    ,dst_vert_grid='L72'
-                                    # ,atm_file=f'{data_root}/HICCUP_TEST.ERA5.atm.low-res.nc'
-                                    # ,sfc_file=f'{data_root}/HICCUP_TEST.ERA5.sfc.low-res.nc'
-                                    ,sstice_name='NOAA'
-                                    ,sst_file=f'{data_root}/HICCUP_TEST.NOAA.sst.nc'
-                                    ,ice_file=f'{data_root}/HICCUP_TEST.NOAA.ice.nc'
-                                    # ,topo_file=f'{data_root}/USGS-gtopo30_ne30np4_16xdel2-PFC-consistentSGH.nc'
-                                    ,grid_dir=data_tmp
-                                    ,map_dir=data_tmp
-                                    ,tmp_dir=data_tmp
-                                    ,verbose=True
-                                    ,check_input_files=False)
+hiccup_data = hiccup.create_hiccup_data(src_data_name='NOAA',
+                                        dst_horz_grid=dst_horz_grid,
+                                        dst_vert_grid='L72',
+                                        sstice_name='NOAA',
+                                        sst_file=f'{data_root}/HICCUP_TEST.NOAA.sst.nc',
+                                        ice_file=f'{data_root}/HICCUP_TEST.NOAA.ice.nc',
+                                        grid_dir=data_tmp,
+                                        map_dir=data_tmp,
+                                        tmp_dir=data_tmp,
+                                        verbose=True,
+                                        check_input_files=False,)
 
 # Print some informative stuff
 print('\n  Input Files')
-# print(f'    input atm files: {hiccup_data.atm_file}')
-# print(f'    input sfc files: {hiccup_data.sfc_file}')
 print(f'    input sst files: {hiccup_data.sst_file}')
 print(f'    input ice files: {hiccup_data.ice_file}')
-# print(f'    input topo file: {hiccup_data.topo_file}')
 print('\n  Output files')
 print(f'    output sst file: {output_sst_file_name}')
 
@@ -79,7 +72,7 @@ print(f'output_sst_file_name: {output_sst_file_name}')
 print()
 
 # Print summary of timer info
-hdc.print_timer_summary()
+hiccup_data.print_timer_summary()
 
 # ------------------------------------------------------------------------------
 # ------------------------------------------------------------------------------
